Add extrinsicsRoot method
