+++ runtime/packages/core-modules/assembly/system.ts
@@ -1,9 +1,9 @@
 import { CompactInt, ByteArray, Bytes, UInt64, UInt32, Hash } from 'as-scale-codec';
 import { Storage } from './storage';
 import { Blocks, Header, ExtrinsicData } from '@as-substrate/models';
-import { Utils } from '@as-substrate/core-utils';
+import { Utils, Serialiser } from '@as-substrate/core-utils';
 import { AccountId } from '@as-substrate/balances-module';
-import { Crypto } from '.';
+import { ext_trie_blake2_256_ordered_root_version_1 } from '.';
 
 export class System {
     // execution phases
@@ -109,7 +109,7 @@ export class System {
     }
 
     /**
-     * Increment extrinsic count
+     * Increment extrinsicCount
      */
     static incExtrinsicCount(): void {
         const count = this.extrinsicCount();
@@ -122,9 +122,22 @@ export class System {
      * @param data 
      */
     static extrinsicsRoot(data: ExtrinsicData): Hash{
-        return Crypto.computeExtrinsicRoot(data.toOrderedTrie());
+        return this.computeExtrinsicRoot(data.toEnumeratedValues());
     }
 
+        /**
+     * Computes the orderedTrieRoot hash of the extrinsics data using 
+     * @param data enumerated values
+     */
+    static computeExtrinsicRoot(data: u8[]): Hash{
+        let dataPtr = data.dataStart;
+        let dataLen = data.length;
+        __retain(dataPtr);
+        const dataU64 = ((dataLen as u64) << 32) | dataPtr;
+        let extrinsicRoot: i32 = ext_trie_blake2_256_ordered_root_version_1(dataU64);
+        const hashU8a = Serialiser.deserialiseInput(extrinsicRoot, 32);
+        return Hash.fromU8a(hashU8a);
+    }
     /**
      * Adds applied extrinsic to the current ExtrinsicData
      * @param ext extrinsic as bytes
