+++ runtime/packages/models/assembly/extrinsic-data.ts
@@ -14,7 +14,7 @@ export class ExtrinsicData{
     toU8a(): u8[]{
         let result: u8[] = [];
         let keys: UInt32[] = this.data.keys();
-        let lenData: CompactInt = new CompactInt(keys.length);
+        let lenData: CompactInt = new CompactInt(<u8>(keys.length));
         result = result.concat(lenData.toU8a());
         for(let i = 0; i < keys.length; i++){
             result = result
@@ -27,14 +27,17 @@ export class ExtrinsicData{
     /**
      * Enumerated items, from which orderedTrieRoot will get 
      * The items consist of a SCALE encoded array containing only values, corresponding 
-     * key of each value is the index of the item in the array, starting at 0. 
-     * The keys are little-endian, fixed-size integers.
+     * key of each value is the index of the item in the array, starting at 0.
+     * In our case same as toU8a() but with only values concatenated
      */
-    toOrderedTrie(): u8[][]{
-        let result: u8[][] = [];
-        let values: ByteArray[] = this.data.values();
-        for(let i=0; i < values.length; i++){
-            result.push(values[i].values);
+    toEnumeratedValues(): u8[]{
+        let result: u8[] = [];
+        let keys: UInt32[] = this.data.keys();
+        let lenData: CompactInt = new CompactInt(<u8>(keys.length));
+        result = result.concat(lenData.toU8a());
+        for(let i=0; i < keys.length; i++){
+            result = result
+            .concat(this.data.get(keys[i]).toU8a());
         }
         return result;
     }
